+++ Darkly/LDEventModel.m
@@ -109,7 +109,7 @@ static NSString * const kCustomEventName = @"custom";
     self.data ? [dictionary setObject:self.data forKey: kDataKey] : nil;
     self.value ? [dictionary setObject:self.value forKey: kFeatureKeyValueServerKey] : nil;
     self.isDefault ? [dictionary setObject:self.isDefault forKey: kIsDefaultServerKey] : nil;
-    self.user ? [dictionary setObject:[self.user dictionaryValueWithFlagConfig:YES includePrivateAttributes:NO config:config] forKey: kUserKey] : nil;
+    self.user ? [dictionary setObject:[self.user dictionaryValueWithFlagConfig:NO includePrivateAttributes:NO config:config] forKey: kUserKey] : nil;
     
     return dictionary;
 }