+++ DarklyTests/LDRequestManagerTest.m
@@ -55,7 +55,7 @@ static const int httpStatusCodeInternalServerError = 500;
     [OHHTTPStubs removeAllStubs];
 }
 
-- (void)testPerformFeatureFlagRequestMakesGetRequestAndCallsDelegateOnSuccess {
+- (void)testPerformFeatureFlagRequest_GetRequest_Success {
     XCTestExpectation *getRequestMade = [self expectationWithDescription:@"feature flag GET request made"];
     
     id requestManagerDelegateMock = [OCMockObject niceMockForProtocol:@protocol(RequestManagerDelegate)];
@@ -76,7 +76,28 @@ static const int httpStatusCodeInternalServerError = 500;
     }];
 }
 
-- (void)testPerformFeatureFlagRequestDoesNotMakeFallbackRequestWhenUsingGet {
+- (void)testPerformFeatureFlagRequest_GetRequest_Success_invalidData {
+    XCTestExpectation *getRequestMade = [self expectationWithDescription:@"feature flag GET request made"];
+
+    id requestManagerDelegateMock = [OCMockObject niceMockForProtocol:@protocol(RequestManagerDelegate)];
+    [[requestManagerDelegateMock expect] processedConfig:NO jsonConfigDictionary:[OCMArg any]];
+    [LDRequestManager sharedInstance].delegate = requestManagerDelegateMock;
+
+    [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest *request) {
+        return [request.URL.host isEqualToString:flagRequestHost] && [request.HTTPMethod isEqualToString:httpMethodGet];
+    } withStubResponse:^OHHTTPStubsResponse*(NSURLRequest *request) {
+        [getRequestMade fulfill];
+        return [OHHTTPStubsResponse responseWithData: [self invalidJsonData] statusCode:httpStatusCodeOk headers:[self headerForStatusCode:httpStatusCodeOk]];
+    }];
+
+    [[LDRequestManager sharedInstance] performFeatureFlagRequest:[LDClient sharedInstance].ldUser];
+
+    [self waitForExpectationsWithTimeout:10 handler:^(NSError * _Nullable error) {
+        [requestManagerDelegateMock verifyWithDelay:1];
+    }];
+}
+
+- (void)testPerformFeatureFlagRequest_GetRequest_DoesNotMakeFallbackRequest {
     NSMutableArray<NSNumber*> *selectedStatusCodes = [NSMutableArray arrayWithArray:[LDClient sharedInstance].ldConfig.flagRetryStatusCodes];
     [selectedStatusCodes addObjectsFromArray:[self selectedNoFallbackStatusCodes]];
     
@@ -119,7 +140,7 @@ static const int httpStatusCodeInternalServerError = 500;
     }
 }
 
-- (void)testPerformFeatureFlagRequestMakesReportRequestAndCallsDelegateOnSuccess {
+- (void)testPerformFeatureFlagRequest_ReportRequest_Success {
     LDConfig *config = [[LDConfig alloc] initWithMobileKey:testMobileKey];
     config.useReport = YES;
     
@@ -129,8 +150,8 @@ static const int httpStatusCodeInternalServerError = 500;
     [[requestManagerDelegateMock expect] processedConfig:YES jsonConfigDictionary:[OCMArg isKindOfClass:[NSDictionary class]]];
     [LDRequestManager sharedInstance].delegate = requestManagerDelegateMock;
 
-    XCTestExpectation *reportRequestMade = [self expectationWithDescription:@"feature flag REPORT request made"];
-    XCTestExpectation *responseArrived = [self expectationWithDescription:@"feature flag response arrived"];
+    __weak XCTestExpectation *reportRequestMade = [self expectationWithDescription:@"feature flag REPORT request made"];
+    __weak XCTestExpectation *responseArrived = [self expectationWithDescription:@"feature flag response arrived"];
     
     [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest *request) {
         return [request.URL.host isEqualToString:flagRequestHost] && [request.HTTPMethod isEqualToString:httpMethodReport];
@@ -151,7 +172,33 @@ static const int httpStatusCodeInternalServerError = 500;
     }];
 }
 
-- (void)testPerformFeatureFlagRequestMakesFallbackGetRequest {
+- (void)testPerformFeatureFlagRequest_ReportRequest_Success_invalidData {
+    __weak XCTestExpectation *reportRequestMade = [self expectationWithDescription:@"feature flag REPORT request made"];
+
+    LDConfig *config = [[LDConfig alloc] initWithMobileKey:testMobileKey];
+    config.useReport = YES;
+
+    self.ldClientMock = [self mockClientWithUser:[self mockUser] config:config];
+
+    id requestManagerDelegateMock = [OCMockObject niceMockForProtocol:@protocol(RequestManagerDelegate)];
+    [[requestManagerDelegateMock expect] processedConfig:NO jsonConfigDictionary:[OCMArg any]];
+    [LDRequestManager sharedInstance].delegate = requestManagerDelegateMock;
+
+    [OHHTTPStubs stubRequestsPassingTest:^BOOL(NSURLRequest *request) {
+        return [request.URL.host isEqualToString:flagRequestHost] && [request.HTTPMethod isEqualToString:httpMethodReport];
+    } withStubResponse:^OHHTTPStubsResponse*(NSURLRequest *request) {
+        [reportRequestMade fulfill];
+        return [OHHTTPStubsResponse responseWithData: [self invalidJsonData] statusCode:httpStatusCodeOk headers:[self headerForStatusCode:httpStatusCodeOk]];
+    }];
+
+    [[LDRequestManager sharedInstance] performFeatureFlagRequest:[LDClient sharedInstance].ldUser];
+
+    [self waitForExpectationsWithTimeout:10 handler:^(NSError * _Nullable error) {
+        [requestManagerDelegateMock verifyWithDelay:1];
+    }];
+}
+
+- (void)testPerformFeatureFlagRequest_ReportRequest_MakesFallbackGetRequest {
     LDConfig *config = [[LDConfig alloc] initWithMobileKey:testMobileKey];
     config.useReport = YES;
     
@@ -214,7 +261,7 @@ static const int httpStatusCodeInternalServerError = 500;
     }
 }
 
-- (void)testPerformFeatureFlagRequestDoesNotMakeFallbackGetRequest {
+- (void)testPerformFeatureFlagRequest_ReportRequest_DoesNotMakeFallbackGetRequest {
     NSArray<NSNumber*> *noFallbackStatusCodes = [self selectedNoFallbackStatusCodes];
     
     for (NSNumber *fallbackStatusCode in noFallbackStatusCodes) {
@@ -465,6 +512,10 @@ static const int httpStatusCodeInternalServerError = 500;
 
 - (NSData*)successJsonData {
     return [[NSData alloc] initWithBase64EncodedString:[LDUtil base64EncodeString:@"{\"test-flag\":true}"] options: 0];
+}
+
+- (NSData*)invalidJsonData {
+    return [[NSData alloc] initWithBase64EncodedString:[LDUtil base64EncodeString:@"invalid json data"] options: 0];
 }
 
 - (NSArray<NSNumber*>*)selectedNoFallbackStatusCodes {