+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the LaunchDarkly iOS SDK will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org).
 
+## [2.10.1] - 2018-02-14
+### Fixed
+- Removes user flag config values from event reports
+- Improves SSE connection error handling
+
 ## [2.10.0] - 2018-02-01
 ### Added
 - Support for specifying [private user attributes](https://docs.launchdarkly.com/docs/private-user-attributes) in order to prevent user attributes from being sent in analytics events back to LaunchDarkly. See the `allUserAttributesPrivate` and `privateUserAttributes` properties of `LDConfig` as well as the `privateAttributes` property of `LDUserBuilder`.